Fix a crash and rename package
