+++ materialchooser/src/main/java/net/theluckycoder/materialchooser/FileItem.kt
@@ -1,3 +1,3 @@
-package net.theluckycoder.filechooser
+package net.theluckycoder.materialchooser
 
 internal class FileItem(val name: String, val path: String, val isFolder: Boolean)
