+++ sample/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 26
     buildToolsVersion "26.0.1"
     defaultConfig {
-        applicationId "net.theluckycoder.filechooser.sample"
+        applicationId "net.theluckycoder.materialchooser.sample"
         minSdkVersion 14
         targetSdkVersion 26
         versionCode 1
